Introduced `ModelReader` and `TensorInfo` in `skainet-io-core` as a common abstraction. - Refactored `MemoryChunk` to support non-copying `slice` views for efficient memory mapping. - Added dependency on `skainet-io-core` in `skainet-io-onnx`. - Fixed alignment parsing and improved tensor data handling in `GGUFReader` tests. - Increased test heap size to 8GB to support large model tests.

+++ skainet-io/skainet-io-gguf/src/commonMain/kotlin/sk/ainet/io/gguf/GGUFReader.kt
@@ -133,7 +133,7 @@ class GGUFReader(source: Source, private val loadTensorData: Boolean = true) {
         offs = newOffs
         val newAlign = fields["general.alignment"]
         if (newAlign != null && newAlign.types == listOf(GGUFValueType.UINT32)) {
-            alignment = newAlign.parts.last()[0] as Int
+            alignment = (newAlign.parts.last()[0] as UInt).toInt()
         }
         val padding = offs % alignment
         if (padding != 0) {
